+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+# 2.0.0 (2020-02-11)
+## Breaking changes
+- Removed `headers/1` in favour of new `req/4` (see below)
+- Bumped required version of Elixir to `~> 1.7`
+
+## New features
+- `req/4` accepts a fourth optional parameter that allows you to send extra
+headers to Stripe, e.g.
+
+```elixir
+%{"Stripe-Account" => "...", "Idempotency-Key" => "..."}
+```
+
 # 1.2.1 (2018-12-05)
 - Relaxed `HTTPoison` dependency version.